Allow GCC 4.7 Compile
Here are many of the changes I found were needed to get this kernel to build with GCC 4.7.  GCC 4.4.3 was too old for my blood and it seemed to work best on 10.04 which I don't use anymore.  GCC 4.7 runs best with current Ubuntu Versions like 12.04 LTS and beyond.

Update cifssmb.c

Signed-off-by: Chester Kener <Cl3Kener@gmail.com>